+++ src/ConcreteEngine.Editor/EditorPortal.cs
@@ -62,9 +62,9 @@ public sealed class EditorPortal : IDisposable
     public void UpdateInput() => ImGuiSystem.FillInput();
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public void UpdateGameTick(float deltaTime) => EditorCamera.Update(deltaTime);
+    public void UpdateGameTick(double deltaTime) => EditorCamera.Update(deltaTime);
 
-    public void Render(float deltaTime, TextureId outputTexture)
+    public void Render(double deltaTime, TextureId outputTexture)
     {
         if (!Initialized) return;
